+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Features
 * Added customizable labels to shape editors.
+* Added the a directive for dragging icons from outside the map over the map while exposing an observables that listen to dragging position.
 
 ## 0.0.33
 ## Fixes